refactor: update user request and change useUser to useAuth
