+++ src/components/NavbarItems/Profile/index.js
@@ -1,21 +1,18 @@
 import React, {useEffect} from "react";
-import styles from "./index.module.css";
 import Link from '@docusaurus/Link';
-import {reloadRequest} from '@site/src/utils/https';
-import {signout} from "@site//src/utils/auth";
-import {useUser} from "@site//src/hooks/useUser";
+import useAuth from "@site/src/hooks/useAuth";
 import clsx from "clsx";
+import styles from "./index.module.css";
 
-// If user has logged in, profile will show avatar, else a login button.
 export const Profile = (props) => {
-    const user = useUser();
+    const {data: user, signOut} = useAuth();
     const isSignIn = user !== null;
 
     if (isSignIn) {
         return (
             <div className={clsx(styles.customNavbarItem, styles.box, {'menu__list-item--hide': props.mobile})}>
                 <img className={styles.avatar} src={user?.user_metadata?.avatar_url}/>
-                <span className={styles.text} onClick={signout}>退出</span>
+                <span className={styles.text} onClick={signOut}>退出</span>
             </div>
         )
     } else {
