+++ src/components/QuizForm/index.js
@@ -1,9 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import Link from '@docusaurus/Link';
 import styles from './styles.module.css';
-import request from '@site/src/utils/https';
+import request from '@site/src/api/request';
 import { SOLIDITY_COURSE_ID } from '@site/src/constants/course';
-import {useLocation} from "@docusaurus/router";
 
 export default function QuizForm(props) {
     const [formUrl, setFormUrl] = useState('#');
