--- src/configs/request.js
@@ -1,5 +1,0 @@
-
-export const BASE_URL = 'https://wtf.arkhodler.com';
-
-
-
