+++ src/hooks/useSession.js
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react"
-import { supabase } from "../utils/auth";
+import { supabase } from "../api/auth";
 
 export const useSession = () => {
 	const [session, setSession] = useState(null);
